requirements.txt
alembic==1.3.2
aniso8601==8.0.0
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-blacklist==0.0.1
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
oauthlib==2.1.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
SQLAlchemy==1.3.12
stripe==2.42.0
urllib3==1.25.7
Werkzeug==0.16.0
